+++ src/bpm.cpp
@@ -4,7 +4,7 @@
 volatile uint32_t ticks = 0;
 float bpm_current = 120.0f; //BPM_MINIMUM; //60.0f;
 //double ms_per_tick = 1000.0f * (60.0f / (double)(bpm_current * (double)PPQN));
-double micros_per_tick = (double)1000000.0 * ((double)60.0 / (double)(bpm_current * (double)PPQN));
+float micros_per_tick = (float)1000000.0 * ((float)60.0 / (float)(bpm_current * (float)PPQN));
 
 volatile long last_processed_tick = -1;
 
@@ -22,7 +22,7 @@ void set_bpm(float new_bpm) {
       set_new_bpm(bpm_current);
     #else
       //ms_per_tick = 1000.0f * (60.0f / (double)(bpm_current * (double)PPQN));
-      micros_per_tick = (double)1000000.0 * ((double)60.0 / (double)(bpm_current * (double)PPQN));
+      micros_per_tick = (float)1000000.0 * ((float)60.0 / (float)(bpm_current * (float)PPQN));
       //Serial.printf("%i bpm is %0.4f beats per second?\n", bpm_current, bpm_current/60.0f);
       //Serial.printf("set ms_per_tick to %f\n", ms_per_tick);
       Serial.printf("set micros_per_tick to %f\n", micros_per_tick);
